+++ lib/src/interop/interop.dart
@@ -1,0 +1,12 @@
+@JS('Msal')
+library msal_js.interop;
+
+import 'package:js/js.dart';
+
+import 'utils/promise_interop.dart';
+
+export 'utils/promise_interop.dart';
+
+part 'logger_interop.dart';
+part 'user_agent_application_interop.dart';
+part 'user_interop.dart';