+++ lib/src/interop/user_interop.dart
@@ -1,0 +1,11 @@
+part of 'interop.dart';
+
+@JS('User')
+class UserJs {
+  external String get displayableId;
+  external String get name;
+  external String get identityProvider;
+  external String get userIdentifier;
+  external dynamic get idToken;
+  external String get sid;
+}