+++ lib/src/user.dart
@@ -1,30 +1,19 @@
-import 'dart:js' as js;
+part of '../msal_js.dart';
 
-/// A wrapper over a JavaScript MSAL user.
 class User {
-  String get displayableId => _handle['displayableId'];
+  String get displayableId => _jsObject.displayableId;
 
-  String get name => _handle['name'];
+  String get name => _jsObject.name;
 
-  String get identityProvider => _handle['identityProvider'];
+  String get identityProvider => _jsObject.identityProvider;
 
-  String get userIdentifier => _handle['userIdentifier'];
+  String get userIdentifier => _jsObject.userIdentifier;
 
-  js.JsObject get idToken => _handle['idToken'];
+  dynamic get idToken => _jsObject.idToken;
 
-  String get sid => _handle['sid'];
+  String get sid => _jsObject.sid;
 
-  /// Gets the Dart wrapper of the underlying User JavaScript object.
-  js.JsObject get jsHandle => _handle;
+  final UserJs _jsObject;
 
-  js.JsObject _handle;
-
-  User._(this._handle);
-
-  /// Creates a wrapper over a JavaScript MSAL user object.
-  /// 
-  /// Note: Intended for internal use only.
-  factory User.fromJsObject(js.JsObject object) {
-    return new User._(object);
-  }
+  User._fromJsObject(this._jsObject);
 }